Add automatic dialogue shrinking based on character image height
This adds an optional new per-stage character attribute, specified in meta.xml as
`<required-image-height>`. This attribute, if present, specifies a
minimum amount of space that must be free of obstruction when displaying
dialogue in game. This attribute is specified in units of source image
pixels (prior to any rescaling).

In general, this corresponds to how tall a character is in their source images, excluding empty space.

If present, `OpponentDisplay` will use this value to automatically calculate
a maximum dialogue bubble height. Dialogue that would overflow this maximum
height will have its `font-size` automatically shrunk to fit the available
space (while still using the largest possible `font-size`).

This also adds logic for automatically shrinking the minimum height for
individual dialogue bubbles, in cases where the default minimum dialogue
bubble size would overlap the character's required image height.
This occurs in tandem with automatic text hiding (if necessary).