Add custom per-character settings
This implements a way for characters to present explicit choices to the
player at the select screen.

Specifically, characters can now define groups of settings in `behaviour.xml`.

Once a character is selected and loaded, the settings within each group
are presented using a dropdown, similarly to the alt costume selection dropdown.

Each group of settings is associated with a marker (which can, as usual,
be either persistent or temporary), and each setting within that group
corresponds to a possible value for that marker.

Each setting can have a custom label within the settings dropdown menu,
as well as an optional set of variable tests that must all match before
the setting is made available to the player.

Finally, a new `settings_changed` trigger has been added, which fires
whenever the player changes the selected option within a group.

===

Internally, settings are represented using `CharacterSettingsGroup`
objects, which themselves contain `CharacterSetting` objects for each
possible setting within a group.

In order to support using variable tests for setting conditions, I've
moved the variable test logic out of `Case.checkConditions` and into
a new `VariableTest` class.