Add syntax for matching characters without a tag
Conditions can now filter for characters without a given tag by prefixing it with `!`.
For instance, `filter="!shy"` will match characters NOT tagged as `shy`.

In particular, this can be used in conjunction with `count="0"` to check if everyone at the table has a tag.

As an example, the condition
```xml
<condition role="aiopp" filter="!shy" count="0" />
```
will match if every opponent (excluding the human player) has the `shy` tag, regardless of how many players are actually present.