Allow use of Saying Marker/Text/Pose targeting Self for post-dialogue hidden cases
This tweaks the case condition checking logic to allow the use of volatile
conditions targeting Self specifically when evaluating post-dialogue hidden cases.

Note that, at least with this basic implementation, Self Saying Marker
does *not* consider other hidden cases; it only checks the dialogue line
that is being displayed on-screen.