I think it's a bit unnecessary to mix suit indices and prefixes, not just because it broke alt texts in the gallery
Also, `ActiveCardImages.prototype.getCardImage` is only called with a `Card`; and why not use a `Deck` to randomize the card backs?

And we don't really need the suit name constants when we use numbers to represent the suits (since ce19d11e9fe1a) and we have them in `SUIT_PREFIXES`. If anything, we need numeric constants.

Speaking of constants, `const` has been a thing for almost ten years; even IE 11 supports it.