Add background image anchoring and overlay options
This adds the ability for background image layers to anchor themselves
to specific areas of the screen, as well as the ability for background
images to overlay character sprites on certain screens.

Image layers can be anchored to the `center`, `left`, or `right` of
the screen. Layers that are anchored to the `center` are constrained to
fit exactly within the 4:3 central area of the screen, beneath the game UI.
On widescreen devices, the remaining space to the sides is filled using
`left`- and `right`-anchored layers (if any).

Layers without a specific anchoring setting span the screen using the
"regular" layout/scaling algorithm for background layers.

Image layers can additionally be specified to overlay the title screen,
the main/group select screens, and the game screen itself. This is primarily
intended to be used with `left`- and `right`- anchored layers to allow
them to frame the game area.

Both of these options can be specified along with the existing `bottom`, `top`, and `repeat` options.