Make cards a player variable, so characters can see how many cards other players swapped
Beware that it will expand to 0 before the player has swapped this round, so like with .hand (which requires psychic powers anyway), only use it in hand lines if the other player is to the left, or combine it with a condition that makes the case volatile.

Also stops fall-through from the `.hand` case.