Some adjustment to the title (character creation) screen
Makes the unselected items easier to make out by relying less on transparancy and more on contrast, as well as adding a colored border to the selected items and options.

Also uses the same `selected` class on the selected gender and size buttons and makes the sizes identical.

Changes the clothing list to a grid layout.

Changes the event handling a bit by moving the handler to spniTitle.js and making use of data attributes.