Fix clothing selection scrolling on mobile
Drops the manual touch scrolling handling, because it works badly with scroll-snap, and instead let touch actions on buttons scroll.

Delegates the event and tooltip handling to the clothing container by associating a reference to the TitleClothingSelectionIcon object to its DOM element with `.data()` to reduce the number of event handlers.

Hides tooltips when scrolling too far, and make sure to hide other tooltips when showing a new one.