Don't remove the hidden.bs.modal handler that's supposed to restore the tabindices from the stripping modal.
Actually, we don't need to attach the event handlers every time the modal is shown and then remove them afterwards.

Instead of focusing the last article of clothing, always remove focus from the active element when showing any modal.

Don't disable spacebar if the active button or input element is invisible or disabled.